rs-app-1.0.0/rs-app-model/src/main/java/es/udc/rs/app/model/domain/State.java
package es.udc.rs.app.model.domain;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.Table;

@Entity
@Table(name = "State")
public class State {
	
	private String id;
	private String name;
	private String description;
	
	// Primary key: id.
	// Not null attribute: name.
	// Unique attribute: name.
	// Business key: id, name. 
	
	public State() {
		
	}

	public State(String id, String name, String description) {
		this.id = id;
		this.name = name;
		this.description = description;
	}
	
	public State(String name, String description) {
		this.name = name;
		this.description = description;
	}

	@Id
	@Column(name = "idState")
	public String getId() {
		return id;
	}

	public void setId(String id) {
		this.id = id;
	}

	@Column(name = "nameState", nullable = true, unique = true)
	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name;
	}

	@Column(name = "descState")
	public String getDescription() {
		return description;
	}

	public void setDescription(String description) {
		this.description = description;
	}

	@Override
	public int hashCode() {
		final int prime = 31;
		int result = 1;
		result = prime * result + ((id == null) ? 0 : id.hashCode());
		result = prime * result + ((name == null) ? 0 : name.hashCode());
		return result;
	}

	@Override
	public boolean equals(Object obj) {
		if (this == obj)
			return true;
		if (obj == null)
			return false;
		if (getClass() != obj.getClass())
			return false;
		State other = (State) obj;
		if (id == null) {
			if (other.id != null)
				return false;
		} else if (!id.equals(other.id))
			return false;
		if (name == null) {
			if (other.name != null)
				return false;
		} else if (!name.equals(other.name))
			return false;
		return true;
	}

	@Override
	public String toString() {
		return "State [id=" + id + ", name=" + name + ", description="
				+ description + "]";
	}
	
}
